Basic Calculator using functions, try and except
This is a basic user input calculator used to perform operations such as Addition, Subtraction, Multiplication and Division. I have tried and used the try and except blocks to the code to avoid the traceback message if the user enters a string instead will give the message programmed by me.